refactor: further splitting and modularization of scripts
Clarify the function of each module
